+++ decoders/src/test/resources/invalid_sns_invalid_s3.json
