+++ decoders/src/test/resources/invalid_sns_valid_s3.json
