+++ decoders/src/test/resources/valid_sns_invalid_s3.json
