+++ decoders/src/test/resources/valid_sns_valid_s3.json
