--- src/main/scala/uk/gov/nationalarchives/aws/utils/SNSUtils.scala
@@ -1,17 +1,0 @@
-package uk.gov.nationalarchives.aws.utils
-
-import software.amazon.awssdk.services.sns.SnsClient
-import software.amazon.awssdk.services.sns.model.{PublishRequest, PublishResponse}
-
-class SNSUtils(snsClient: SnsClient) {
-  def publish(message: String, topicArn: String): PublishResponse = {
-    snsClient.publish(PublishRequest.builder()
-      .message(message)
-      .topicArn(topicArn)
-      .build())
-  }
-}
-
-object SNSUtils {
-  def apply(snsClient: SnsClient): SNSUtils = new SNSUtils(snsClient)
-}
